docker_build.py qof improvements

+++ docker_build.py
@@ -1,6 +1,9 @@
 import os
 import yaml
 import subprocess
+import logging
+
+logging.basicConfig(level=logging.INFO, format='[%(asctime)s] - %(message)s', datefmt='%H:%M:%S')
 
 
 def cleanup_untagged_images():
@@ -10,9 +13,9 @@ def cleanup_untagged_images():
     remove_command = ["docker", "image", "prune", "-f"]
     result = subprocess.run(remove_command, capture_output=True, text=True)
     if result.returncode == 0:
-        print("[docker_cleanup]: Successfully removed untagged images.")
+        logging.info("Successfully removed untagged images")
     else:
-        print(f"[docker_cleanup]: Error removing untagged images: {result.stderr}")
+        logging.error(f"Error removing untagged images: {result.stderr}")
 
 
 def generate_dockerfile(component, template_path, base_dir_path):
@@ -23,39 +26,36 @@ def generate_dockerfile(component, template_path, base_dir_path):
     :param template_path: path to the dockerfile.template
     :param base_path: base path where component directories are located
     """
-    # Create the needed paths for the component
     component_path = os.path.join(base_dir_path, component)
     dockerfile_path = os.path.join(component_path, 'Dockerfile')
+    if not os.path.exists(template_path):
+        logging.error(f"Template path '{template_path}' does not exist")
+        return
 
-    # Read the Dockerfile template and write to the component's directory
     with open(template_path, 'r') as template_file:
         template = template_file.read()
     with open(dockerfile_path, 'w') as dockerfile:
         dockerfile.write(template)
-
-    # Build the Docker image
     tag = f"myapp/{component}:latest"
     build_command = ["docker", "build", "-t", tag, "."]
     result = subprocess.run(build_command, cwd=component_path, capture_output=True, text=True)
-
     if result.returncode == 0:
-        print(f"[docker_generate]: Docker image for {component} built successfully.")
+        logging.info(f"Docker image for {component} built successfully")
     else:
-        print(f"[docker_generate]: Failed to build Docker image for {component}: {result.stderr}")
+        logging.error(f"Failed to build Docker image for {component}: {result.stderr}")
+
+
+def main(base_dir_path: str, template_path: str):
+    logging.info("Welcome to the Docker image generator, application starting...\n")
+    # Base the container creation on the components defined in the dag
+    with open('application_dag.yaml', 'r') as dag_file:
+        app_dag = yaml.safe_load(dag_file)
+    for component in app_dag['System']['components']:
+        generate_dockerfile(component, template_path, base_dir_path)
+    cleanup_untagged_images()
 
 
 if __name__ == '__main__':
-    # Load the application DAG from the YAML file
-    with open('application_dag.yaml', 'r') as dag_file:
-        app_dag = yaml.safe_load(dag_file)
-
-    # Define the paths
     base_dir_path = 'components'
     template_path = 'src/dockerfile.template'
-
-    print(f"Welcome to docker image generator, application starting...\n")
-    # Generate Dockerfiles for each component
-    for component in app_dag['System']['components']:
-        generate_dockerfile(component, template_path, base_dir_path)
-    # Remove previous unused images
-    cleanup_untagged_images()
+    main(base_dir_path, template_path)
